Fill out the rest of the assembler proof procedure
This completes the first stage of the verification proof.
In the assembly stage, we prove a theorem like:

is_asmp foo_start "0x0F 0xB0 ..." foo_start foo_end (
  asmp_at 0 ( -- entry block
    -- NEG rax
    (instUnop unopNeg x0) +A
    -- ADD rdx, 37  ; add = 0, rdx = 2
    (instBinop x0 x2 (IRM_imm32 37)) +A
    -- JUMP to block 2, which is at byte position 7 (say)
    (instJump 7)
  ) +A
  asmp_at 7 ( -- block 2
    -- CALL bar, which is at bar_start
    (instCall bar_start)
    -- ret
    (instRet)
  )
)

where `+A` is the `asmp_A` sequencing operator. The important properties
of this representation are:

* Individual blocks are labeled by their byte position relative to the
  procedure start.

* All internal jumps (like `instJump 7` in the example) are also
  resolved to procedure-relative addresses instead of RIP-relative
  (which changes often since RIP is different for every instruction).

* Procedure calls are resolved to global addresses instead of
  RIP-relative, again so that we can store the list of all the functions
  and their start positions and resolve where a call goes to without
  having to keep fine grained track of RIP.

* More obviously, the representation is not bytes anymore.
  Working with bytes is fiddly and also chains us to the intel spec.
  By translating to a more logical notion of instruction, we can
  abstract away from details like XOR clear vs writing 0, as long as the
  instructions have similar enough execution semantics.

* We can also handle pseudo-instructions, although right now we aren't
  taking much advantage of this ability except for the assert
  pseudo-instruction `jump +2; ud2`; it is convenient to not need
  another block just for this.

The second phase of verification, where most of the interesting logic
goes, is the proof of correctness of the procedures, in dependency order
(i.e. leaf functions first, then working backward to main).
The representation here sets us up to be able to do random access of the
procedures and know where we are at any given point.

There is an interstitial step between the two phases, which takes the
accumulated list of assembled procedures, and breaks it up into
components (one per procedure) and puts them in separate theorems.
That way we don't have to pay for an access to the global list every
time we access a function definition. This step also handles back
propagation of the `ip e. u64` assumption in `parseInst` (which is there
so that we don't have to constantly prove that the instruction pointer
is in bounds and can do it only once at the end).