Add assembled procedure theorems
This does the lemma-making step I talked about in the previous commit.
Since the bad behavior only shows up when we have a large number
of lemmas to split out, it bottoms out to the naive method when there
are at most 16 conjuncts, which is already pretty large. This means that
approximately n/16 lemmas will be created for large n, which seems like
reasonable overhead, and for small examples you won't need any lemmas at
all.