More compiler lemma design
The main contribution is some more theorem statements for the compiler.
The `compiler.lean` file is a sketch of the overall structure including
types; the representation of some of the core types like `tctx` are
still undetermined, but the hierarchy of contexts `gctx -> pctx -> bctx`
seems to work well for theorems like `okInst_loop`.

Unfortunately, I can already tell that the theorems are going to be very
large and unreadable in the usual display style. I am taking advantage
of the fact that reusing a large term is cheap in MM0 as long as you
don't pattern match on the term or inspect its contents. For example,
`lctx` contains a hypothetical with all the label groups corresponding
to back-edges in the control flow graph, but this hypothetical has to
carry the type context for the targets, and these type contexts are both
large and share a lot of subterms with each other (because the type
contexts all derive from a similar root). This information is all placed
left of the `->`, where it appears in every step while all the
interesting stuff happens in the `P` and `Q` arguments to `okInst` and
`okInsts`. The reason for the nested contexts is to put the things that
vary the most near the root, so we don't have to pattern match any more
than necessary.

Without MM0's performance characteristics, this would be a very risky
proof strategy, with potential for something like O(n^3) blowup in the
worst case. And it's still problematic for docgen, which does not
deduplicate subterms in expressions. But this is the most efficient way
to utilize the computation-by-matching architecture to get something
roughly equivalent to a functional programming language.

One way to mitigate the issue on docgen's end is to construct ad-hoc
abbreviations for subterms that are reused many times. The tricky part
is determining which subexpressions to abbreviate, and how to expand
the abbreviations on demand. I think this strategy will do well on
Metamath and MM0 proofs, and it will help lessen the impression that
the proofs are worse than they actually are.

Minor changes:

* The `arch/x86/proof.rs` file ended up not being used, so it has been
  removed.
* Fixed a bug with `VInst::Assert` not having a target.
* Added `PInst::Fallthrough`, plus support for ghost instructions in
  proof generation.
* `PInst::Load64/Store` now indicate if they are spills from regalloc.
  `MovRR` steps are always regalloc insertions, so it is now possible
  to tell by looking at an instruction whether it is a regalloc
  insertion or not.