Implement caching using bincode + serde
This is a relatively low effort approach to caching the results of
elaboration in a `.mm1.o` file. I'm putting this on a branch for
posterity, but I'm going to abandon this approach because both
serialization and deserialization are very slow; it is slower than no
caching at all, at least for compiler.mm1. There are also some
correctness issues around some of the more obscure kinds of lisp data.

I believe the issue is due to lack of proper serde support for Rc.
MM0 relies a lot on dag-like sharing, and while this is explicit in the
core calculus `Expr` and `ExprNode` types, in `LispVal` this sharing is
represented using `Rc`, and serialization and deserialization both make
no attempt to deduplicate `Rc`'s, resulting in a potential exponential
blowup. Breaking reference equality also breaks `LispRef`, so user data
will probably be slightly broken anyway.

The correct way to do this is as an add-on to the MMB format, with a
custom serialization for lisp data and relying on the main formal
content for everything else. It is already possible to do this to some
extent, if you generate a `.mmb` file and then import it instead of the
`.mm1` file that produced it. This is very fast but does not include
lisp data so you only get the raw theorems and not any associated
automation.