Working on block tree
This implements more of the block tree handling, but I think I will
switch to another representation for the block tree, also implemented
here.

Rather than keeping the block tree structure separate from the blocks,
we can instead represent `Many` nodes (which represent forward jumps)
as `DominatedBlock` statements. This branches the context at the current
point to declare a statement that will be jumped to later. For example,
an if statement looks like:

```
  dominated_block(after);
  if cond {h. jump tru(h)} else {h. jump fal(h)};

tru(h: cond):
  ...
  jump after; // predeclared above

fal(h: !cond):
  ...
  jump after; // predeclared above

after:
  // `jump after` is not valid after this point,
  // since dominated_block() only allows forward jumps
  ...
```

Backward jumps and the `Induction` node are handled via the
`LabelGroup` statement. For example, with a while loop:

```
  dominated_block(after);  // allow forward jumps to after
  label_group([base]);     // allow backward jumps to base
  jump base;               // considered a backward jump

base:
  if cond {h. jump main(h)} else {h. jump fal(h)};

main(h: cond):
  ...
  jump base;  // ok because we are in scope of label_group([base])

fal(h: !cond):
  jump after;  // ok because we are in scope of dominated_block(after)

after:
  // `jump base` or `jump after` is not valid at this point,
  // since dominated_block() only allows forward jumps
  // and the label_group([base]) declaration is out of scope
  // (because dominated_block(after) was declared first)
  ...
```

This scheme lets us represent the tree structure inline to the CFG
itself, which also means that we don't have to reconstruct the whole
context, and the context of `after` can just be a diff from the context
at the point of the `dominated_block(after)` statement.