Add variable context management
The variable context is one of the key data structures in the compiler,
so I thought it would be good to handle it efficiently even though it's
probably overkill for the present.

Like everything in MM0, variable contexts have to be stored as binary
trees, but contexts go through a lot of copying and branching, as well
as significant amount of unshared manipulation. We'd really like to
avoid quadratic time for variable context manipulation, even in the face
of long basic blocks declaring many variables.

To address this, we use a representation like:

vctx := (((A4 *V A2) *V B8) *V B2) *V B1

where in this case we have a context B with 11 variables,
added to a shared context A with 6 variables
(so in total 17 variables are in scope).
A4 represents a perfect binary tree of size 4 containing the first 4
variables, and similarly for the others.

This representation has the advantage that all operations are worst case
logarithmic (except for deep sharing, where there are many small
shared contexts), and it also achieves sharing of context A with other
variable contexts. For unshared contexts, we have a mutation bonus:
the push operation on this representation is amortized O(1).