Add a "small trusted kernel" to mm0-rs
The mm0-rs proof assistant is fundamentally not part of the trusted
base, so there is no need for it to have a trusted kernel since the
kernel is external (mm0-c). However, it is still good to be able to
simply cross check proofs, especially those that are generated directly
by tactics (either user tactics like `norm_num` or builtin tactics like
the WIP compiler tactic), as these skip the normal elaborator pipeline,
which is the first line of defense against invalid proofs.

Once a bad proof is in the elaborator state, it is not easy to get it
rechecked even if you know there is an issue; and mm0-c error messages
are essentially stack dumps so there is a need for something with a
focus on better error messages.

This verifier component doesn't really deliver on better error messages
currently, but at least it checks all proofs as they are entered into
the elaborator so it's a bit easier to notice and track down bad proofs.