implement bytecode VM for MM1
This converts the core evaluator for the scheme-like MM1 metaprogramming
language from a state machine (with complex State and Stack types)
to a stack machine and bytecode evaluator, more similar to the core of
languages like lean 3 or python.

* Each Ir instruction is now a stack machine operation, and code is
  a sequence of Ir instructions instead of a tree.
* The Stack values are mostly not driving computation now; each Ir
  knows what Stack values it needs to read or write.
* The Pattern type is merged with Ir, for better space usage. With only
  a few exceptions, Pattern instructions only evaluate in the
  pattern_match() function, and Ir instructions evaluate in run().
* `Refine` is still a state machine; it might be converted to a stack
  machine later if it is worth the performance.

Testing this commit shows performance about the same, slightly better
than the original; but there are a lot more opportunities for
improvement in this code, and hopefully it can be performance optimized
a bit better.