splitting up Evaluator.run()
Just some minor cleanup, mainly extracting rarely used instructions into
their own functions. I was hesitant to do this because this is the
hottest function in the program, but doing it for rare instructions is
good for instruction cache locality. It also turns out to be good for
compilation - this compiles noticeably faster in release mode
(4m35s -> 3m58s for a full rebuild), presumably because LLVM has
superlinear behavior on large functions.