add procedure calls (the empty function compiles!)
This is a pretty weak milestone, especially since there are still
both explicit `sorry` applications and tons of theorems being proved by
`sorry`. The latter issue won't be addressed for a while, because the
shape of the theorems affects the definition of the predicates,
which affects the proofs. Still... :tada: