add axiom sets to docgen
This helps to cut down on the list of axioms in displayed theorems
(especially `axs-string` which has 23 axioms, all of which are
conservative over PA and 16 of which are very similar, which makes it
harder to review the list at a glance).
All axioms in peano.mm1 have been categorized, except for `sorry` for
obvious reasons.