okProc / prologue refactor
The way data gets passed from the procedure entry to the prologue
sequence has changed, summarized in the type of okProcI:

theorem okProcI:
  $ okAssembled (mkPCtx gctx (mkPCtx1 ret epi se)) (asmEntry start code) $ >
  $ accumArgs args vctx1 sz1 $ >
  $ accumClob clob mctx1 mctx2 $ >
  $ okCode
    (mkBCtx (mkPCtx gctx (mkPCtx1 ret epi se)) LG0)
    (okPrologue epiRet x0 mctx2 (mkPrologue vctx1 sz1 epi))
    code
    ok0 $ >
  $ okProc gctx start (mkArgs args mctx1) ret clob se $;

We parse the arguments into (vctx1, sz1), then use the argument machine
context mctx1 and parse the clobbers (which extends the machine context
with `clob |-> FREE`) to mctx2, and then pass all of this to the
prologue, which adds a clobbered register for each caller-saved reg.
None of the additions to mctx are checked at this point, it just
accumulates the registers from these three sources. At the end of the
prologue we have one of:

theorem okPrologue_alloc:
  $ okMCtx mctx $ >
  $ sp + n = m $ >
  $ m < 2^12 $ >
  $ okCode bctx
    (okPrologue (epiFree n epi) sp mctx (mkPrologue vctx sz epi))
    (instBinop x5 wSz64 x4 (IRM_imm32 n))
    (mkTCtx vctx sz (mctx +M stkFREE n)) $;

theorem okPrologue_alloc0:
  $ okMCtx mctx $ >
  $ okCode bctx (mkTCtx vctx sz mctx) code lctx2 $ >
  $ okCode bctx (okPrologue epi sp mctx (mkPrologue vctx sz epi)) code lctx2 $;

depending on whether we need stack space or not. This is where we build
the actual TCtx from the vctx and mctx pieces, and also check that mctx
is ok, which here means that it consists only of distinct register
assignments. As a result, the TCtx can assume that mctx is a proper
separating conjunction.