add instruction classification pass
The basic idea is to augment the build_vcode pass with an auxiliary data
structure that describes enough information to reconstruct, for each
x86 instruction, which MIR level instruction and what part of the
instruction was involved in the creation of that x86 instruction.
This is more reliable than attempting to reverse engineer the purpose of
each instruction based on minimal ghost instructions and replication of
the build_vcode logic inside proof::compiler.