arena allocation for proofs
This decreases memory consumption and speeds things up a bit,
at the cost of a bit more indirection when walking proof objects.

The main change is that ExprNode and ProofNode are no longer directly
recursive types, and instead they store `usize` pointers into a `store`
vector which can be thought of as a local bump allocator. All subterm
data for a given node is contiguous in the vector, so that it can be
referred to by a single pointer, although the node itself might be
stored elsewhere like in `hyps` or `ret` if it's not a subterm of
something else.

We now have to thread these `store` variables everywhere, so a bunch of
functions need to take more arguments. Besides this, `ProofNode::Term`
nodes no longer store their length inline (implicitly as a result of
holding a `Box<[ProofNode]>`); the arity of a term must be looked up in
the environment. This makes proofs smaller at the cost of having to do
a lookup for each App/Term/Thm/Cong node, although in almost all
situations we were already doing the lookup for one reason or another.