change `<>` pair notation to `,`
This is just an objectively good change, and it makes me very happy to
not have to explain this notation whenever it comes up. I had convinced
myself that it wouldn't work because `,` is also used in several other
notations, but it basically just works, and the entire change is just a
big find-replace job.

I kept the precedence as is to keep the diff small, and also because I
think the current precedence is a good choice for not needing too many
parentheses, but I do notice that `,` seems to work naturally with a
lower precedence. It is already lower than most binops, and it is nice
that `a, b e. c` works, so I will keep it as is and see if I can get
used to it.