add superfluous parenthesis linter
Because it has a high false positive rate, it is not enabled by default,
but it was quite helpful in finding unnecessary parentheses in the
existing files. (About half of the examples that trigger the warning
were deliberately left unchanged, because of examples like
`(a <-> b) <-> (c <-> d)` which are made more confusing with minimal
parentheses.)

Properly handling the edge cases most likely requires the partial order
precedence system that was considered and rejected in favor of the
current simple linear order approach. A partial order would allow us to
detect "confusables" like `(a -> b) <-> c` and suppress the lint in
those cases, even though there is technically a precedence between
these two.

Cases like `(a + b) + c = a + (b + c)` can't really be detected by this
mechanism though; we want to allow `a + b + c` because it is an
associative operator, but in the statement of associativity itself it is
often clearer to keep things explicitly parenthesized.

In the end, it's a judgment call, so I'm just locally enabling the
warning and fixing obvious improvements and ignoring things that seem
deliberate.

To use the new linter, just use `(warn-unnecessary-parens #t)` in any
file, or `--warn-unnecessary-parens` on the command line.