fix &T semantics
I previously said that `&'a T` is equivalent to `own(ref 'a T)`, since:

  [x: own T] := E. v: T, x |-> v
  [x: &'a T] := E. v: ref a T, x |-> v

But this is not correct. This would imply that having [x: &'a T] would
include an exclusive points-to assertion x |-> v for some v, which is
not the case for a shared reference. The correct expression just says
that [x: &sn v] instead of x |-> v, which can be written `x = &v` using
the new "address-of" pure expression operator in the PDF. That is:

  &'a T := E. v: ref a T, &sn v
  [x: &'a T] := E. v: ref a T, x = &v

This is not equivalent to `own` of anything, so we have to adjust the
desugaring.

Interestingly, pattern matching an `own T` and a `&'a T` acts the same
way: We end up with `v: T` and `x: &sn v` either way. The difference is
in the (non-user-visible) place where `v` is stored: in the `own T` case
`v` is stored somewhere on the heap, and the type context takes
ownership of the `x |-> v` part of the proposition, while in the `&'a T`
case we get `v: T` and `x: &sn v` directly out of the proposition but
`v` is in external memory associated to the lifetime `'a`.

To accommodate this, the `Own` existential pattern matching statement
has been renamed to `Ptr` since it handles both `own T` and `&'a T`
matching.