assign to returns even after noreturn function
This should probably be done another way, but the issue is that it is
possible for a MIR function to call a function with a generic return
type and deduce that the code after the function is "reachable", but
later in build_vcode we find that it is actually not reachable because
the return type is uninhabited. In that case, our analysis of reachable
blocks is incorrect in visit_blocks() - we can visit blocks even though
we also think they are unreachable, and this left-hand-right-hand
problem causes the register allocator to get a messed up state in these
blocks.

To fix the issue, we now always assign the return variables even if we
think that the function is noreturn, which results in dead code being
emitted which is not great but better than a crash. The better solution
would be to make sure the reachability analysis is precise.