track block parents
Every block has a parent. The two blocks targeted by a branch have the
source of the branch as parent, calls and jump1 have a unique
predecessor, and for general jumps, the block is "declared" by a
DominatedBlock command in the parent. The context as of this instruction
in the parent block determines most of the variables in the child block,
except for variables defined on the spot (e.g. the new hypothesis
variable in a branch or the block parameters in a label).

This makes each block keep track of not only its full context, but also
the prefix of that context that represents the context of the parent
(so the diff between them is the set of variables declared in the block
itself). Besides looking less overwhelming in printouts, this solves an
O(n^2) problem where a sequence of blocks get ever larger contexts:

bb0():
  let x1: bool := true;
  jump1 -> bb1;
bb1(x1: bool):
  let x2: nat := 1;
  jump1 -> bb2;
bb2(x1: bool, x2: nat):
  let x3: nat := 1;
  if x1 then h. bb3 else h. bb4;
bb3(x1: bool, x2: nat, x3: nat, h: x1):
  ...
bb4(x1: bool, x2: nat, x3: nat, h: !x1):
  ...

These are now represented as:
bb0():
  let x1: bool := true;
  bb0+1 := jump1 -> bb1;
bb1(bb0+1):
  let x2: nat := 1;
  bb1+1 := jump1 -> bb2;
bb2(bb1+1):
  let x3: nat := 1;
  bb2+1 := if x1 then h. bb3 else h. bb4;
bb3(bb2+1, h: x1):
  ...
bb4(bb2+1, h: !x1):
  ...

By representing the parent context implicitly, the context of each new
block stays constant size both in the printout and in processing.