[axiom update] make input count up instead of down
Now the `ksIn` field of the model grows rather than shrinks when you
read something from input. This is a bit weird if we think of it as a
deterministic program which is reading the input from a feed, but this
works better for stating properties about the final input-output
relation without having to have memory of the initial state.

One thing that this reformulation misses is the ability to tell when
the input is exhausted. The previous version was axiomatized to never
return an empty string from `read()` if there are remaining bytes,
but this version cannot express that constraint, and it is possible to
have `read()` spuriously return the empty string, at which point the
program will think the input is finished (although if it calls the
function again it might get more data), and it will successfully exit
with a prefix of the actual input.

A more accurate model would take the input to have a boolean "is closed"
flag which is enabled when the input runs out; then we could promise
that the input is not closed except at an empty read(). This would
complicate the type of terminates_ensuring though since programs are
allowed to exit before exhausting the input. Maybe revisit?