Use local addresses for procedures
Now procedure calls are relative to `text_start` instead of being
absolute addresses (and local labels are still relative to the procedure
start). This makes the numbers smaller; it turns out that we don't ever
need the `text_start` offset, at least for calls; since it's a constant
we can add it back in if it comes up (e.g. constant references).

assembledI is now closer to the assumptions in the full ELF proof.
I considered skipping the `filesz <= memsz` check, because it is implied
by later checks in the compiler proof, but that's a bit overkill to
delay a simple inequality so long, and it is nice to be able to get a
`isBasicElf` proof just from the assembler part.