[axiom update] x86 instruction max length
The Intel spec says that x86 instructions must be at most 15 bytes.
This is not true by a naive reading of the encodings; it is possible to
combine enough prefixes and long encodings to create an instruction that
would be longer than 15 bytes, but such an instruction will cause an
illegal instruction exception.

Spec-wise, we can encode this constraint as an extra conjunct in
`decode`; but for the specific instructions supported by the spec the
maximum possible length is only 12, so the conjunct is superfluous
and we prove this.