simplify Proof to EnvFull
Instead of defining a single proof object that covers the entire
environment, it is possible to just assert separately that each theorem
is provable.  (We can't use the same trick for theorem proofs because
proofs are not structurally inductive over the theorem statement.)
This is a lot simpler since we don't need to keep track of a whole set
of possible proofs and prove that they are bounded by something.