don't put spaces before `,` as an infix operation
Not sure how best to do this generically. I'm kind of okay with the
hardcoded names here, since this is really just a list of symbols
that lend themselves well to being displayed without a space before
as in `a, b` instead of `a , b`. The spacing was originally introduced
because just because the symbol has a delimiter doesn't mean it
should be no-spaced. For example if `-` is a left delimiter we probably
don't want to print `a - b` as `a -b` even though it is not ambiguous.
But comma looks really bad with spaces on both sides like this.