add support for math comments
Comments inside math are not supported in MM0, but it was accidentally
working after `)` due to reuse of the outer MM0 parser. This fixes the
bug for MM0 mode parsing, and also makes comments in math work properly
in MM1 mode, including semantic highlighting. (This does mean that you
can't use `--` as a math token in MM1. I don't think it is worth it on
balance.)