add unused hypothesis linter
This one is also categorized under the same `unused-vars` check. It's
pretty easy to implement since we are already keeping track of whether
a subterm is used once or more than once in `Dedup`, and since one
reference corresponds to the hypothesis array itself, this tells us
whether the hypothesis is used at all.

And of course, as soon as we deploy a new linter we find a half dozen
violators in the wild. `assembler-new.mm1` is a special case since every
proof by `sorry` apparently doesn't use any of the hypotheses. It is
possible to get around this by using `'(sorry h1 h2)` instead of
`'sorry`, but this is needless busywork that will make grepping for
`'sorry` more complicated, so instead I just turned off the linter in
that file.