add support for rename request
The rename request itself was relatively easy since it's essentially the
same as the references request; we have to distinguish between
an appearance of `iff a b` and `a <-> b` now though;these act the same
in the references request but a rename of `iff` should not target
notations.

The harder part was the prepareRename request, because without it
vscode includes the `:` when you try to rename the `h` in
`(h: $ foo $)`, because `:` is a "word character" (which is true, in
lisp mode but not in the outer syntax). So we have to manually resolve
the right span here.