fix terminates_ensuring
The previous definition asserts that every reachable state *can* reach
an exit, which means that in a state graph like 0 -> 0 -> 1 where 1 is
an exit state, 0 is considered terminating, even though there is an
infinite path 0 -> 0 -> ... in the graph. The new definition fixes this
by using an inductive predicate.

There is an interesting complication here that PA cannot actually
define inductive predicates such as this in general. This is a minor
variation on the accessibility relation `Acc`. Luckily, we are saved
because the relation `ksStep` satisfies `finite {k2 | ksStep k k2}`,
so `k` is accessible if and only if there is a uniform bound on the
length of paths from `k`, or equivalently the set of all paths
from `k` is finite.